input

+++ src/main/java/com/project/warehouse/dto/OutputProductDto.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 @Data
 public class OutputProductDto implements Serializable {
     private final Long productId;
-    private final BigDecimal amount;
-    private final BigDecimal price;
+    private final Double amount;
+    private final Double price;
     private final Long outputId;
 }
